move mouse handling functions to own module
